testing/web-platform/tests/atomics/dedicated-worker-agent/wait/did-timeout.html

<!doctype html>
<meta charset=utf-8>
<title></title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/resources/workerhelper.js></script>
<script>

// Copyright (C) 2015 Mozilla Corporation.  All rights reserved.
// This code is governed by the BSD license found in the LICENSE file.

/*---
description: >
  Test that Atomics.wait returns the right result when it timed out and that
  the time to time out is reasonable.
---*/

window.addEventListener("load", function() {
  async_test(function(t) {
    var worker_code = `
      onmessage = function(e) {
        var ia = e.data[0];
        var then = Date.now();
        var ret = Atomics.wait(ia, 0, 0, 500); // Timeout 500ms
        postMessage([ret, Date.now() - then]);
      }
    `;
    var onmessage = this.step_func(function(e) {
      var ret = e.data[0];
      var time = Math.abs(e.data[1]);
      assert_equals(ret, "timed-out");
      assert_true(time >= 500);
      t.done();
    });
    var ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
    var worker = WorkerHelper.createFromString(worker_code, onmessage);
    worker.postMessage([ia]);
  }, "atomics-dedicated-worker-wait-did-timeout");
});
</script>
